// packages/cloud/src/models/transition.ts
import { z } from 'zod';
import { WorkflowRulesSchema } from '#/models/workflowRules';
import { TransitionScreenDetailsSchema } from '#/models/transitionScreenDetails';

/** Details of a workflow transition. */
export const TransitionSchema = z.object({
  /** The description of the transition. */
  description: z.string(),
  /** The statuses the transition can start from. */
  from: z.array(z.string()),
  /** The ID of the transition. */
  id: z.string(),
  /** The name of the transition. */
  name: z.string(),
  /** The properties of the transition. */
  properties: z.record(z.string(), z.any()).optional(),
  rules: WorkflowRulesSchema.optional(),
  screen: TransitionScreenDetailsSchema.optional(),
  /** The status the transition goes to. */
  to: z.string(),
  /** The type of the transition. */
  type: z.enum(['global', 'initial', 'directed']),
});

export type Transition = z.infer<typeof TransitionSchema>;